requirements.txt
prisma
asyncio
python-dotenv
requests
logging
typing
pyright

